支持多目录，支持登录认证

+++ src/library/HomeApi.php
@@ -6,9 +6,15 @@ namespace yiqiniu\logger\library;
 
 use think\App;
 use think\Template;
+use yiqiniu\logger\library\traits\AuthTrait;
 
+/**
+ * Class HomeApi
+ * @package yiqiniu\logger\library
+ */
 class HomeApi
 {
+    use AuthTrait;
 
     protected static $_instance = null;
     protected $app;
@@ -20,6 +26,7 @@ class HomeApi
         'tpl_cache' => false,
         'tpl_replace_string' => [],
     ];
+
 
     /**
      * HomeApi constructor.
@@ -50,12 +57,62 @@ class HomeApi
         return self::$_instance;
     }
 
+
+    /**
+     * 登录验证
+     */
+    public function login()
+    {
+        // 设置模板引擎参数
+        $template = new Template($this->tpl_config);
+        // 读取模板文件渲染输出
+        if ($this->checkLogin()) {
+            return redirect('/viewer/index');
+        }
+        $template->fetch('login');
+    }
+
+
+    /**
+     * 认证
+     */
+    public function auth()
+    {
+        $post = input('post.');
+        if (empty($post['username']) || empty($post['password'])) {
+            return json(["code" => 1, "msg" => '用户名或密码不能为空']);
+        }
+        $password = $this->getPassword($post['password']);
+        if (!$this->checkUserinfo($post['username'], $password)) {
+            return json(["code" => 1, "msg" => '用户名或密码错误']);
+        }
+        return json(["code" => 0, "msg" => '登录成功', 'url' => url('/logger/index')]);
+    }
+
+
+    /**
+     * 退出登录
+     * @return \think\response\Json
+     */
+    public function logout()
+    {
+        cookie('logger_info', null);
+        return redirect('/viewer/login');
+    }
+
+    /**
+     * 显示内容
+     */
     public function index()
     {
         // 设置模板引擎参数
         $template = new Template($this->tpl_config);
-        // 模板变量赋值
-        //$template->assign(['root' => $root_list]);
+        if (!$this->checkLogin()) {
+            $template->fetch('login');
+            return;
+        }
+
+
         // 读取模板文件渲染输出
         $template->fetch('index');
     }
@@ -65,6 +122,9 @@ class HomeApi
      */
     public function treelist()
     {
+        if (!$this->checkLogin()) {
+            return redirect('/viewer/login');
+        }
         $list = Logger::getInstance($this->app)->treelist();
         return $this->result($list, count($list));
     }
@@ -74,6 +134,9 @@ class HomeApi
      */
     public function filelist()
     {
+        if (!$this->checkLogin()) {
+            return redirect('/viewer/login');
+        }
         $path = input('get.path', '');
         $list = Logger::getInstance($this->app)->filelist($path);
 
@@ -85,9 +148,14 @@ class HomeApi
      */
     public function filecontent()
     {
+        if (!$this->checkLogin()) {
+            return redirect('/viewer/login');
+        }
         $path = input('get.path');
         $content = Logger::getInstance($this->app)->read($path);
-        return '<div style="padding: 5px; font-size: 16px; line-height: 25px;white-space:nowrap;">'.str_replace(["\r\n","\r","\n"],'<br/>',$content).'</div>';
+        $prefix = '<div style="padding: 5px; font-size: 16px; line-height: 25px;white-space:nowrap;">';
+        $suffix = '</div>';
+        return $prefix . str_replace(["\r\n", "\r", "\n"], '<br/>', htmlspecialchars($content)) . $suffix;
     }
 
     /**
@@ -95,6 +163,9 @@ class HomeApi
      */
     public function delete()
     {
+        if (!$this->checkLogin()) {
+            return redirect('/viewer/login');
+        }
         $path = input('post.path');
         $result = Logger::getInstance($this->app)->delete($path);
         return $this->result([], 0, '删除完成');
